Implement simple button and LED handling
This code implements button press and LED handling with MINAR.

Out of reset, the device boots into config mode, then after 30s
transitions to beacon mode.

Pressing the button will stop all advertising, from either mode,
and put the beacon to sleep. A 1s flash of the LED defined as
red_led in the config.json will convey that the beacon is now 'off'

From the off state, pressing the button will restart the beacon in
config mode, as if it has been powered on. This will be indicated
by a 1s flash of the main led, 'led' in config.json

The included config.json is for a MinewTech i7 beacon